real-time-speech-denoising-nonstationary-noise/requirements.txt
numpy==1.26.4
scipy==1.11.4
soundfile==0.12.1
pyyaml==6.0.1
tqdm==4.66.1
pystoi==0.4.1
# PESQ can be platform-specific. If install fails, set compute_pesq: false in configs/eval_default.yaml
pesq==0.0.4
